+++ CHANGELOG.md
@@ -223,4 +223,4 @@
 
 * Added method to get list of supported DEX's
 * Added method to change DEX's
-
+* Added method to get list of supported tokens